refactor: vehicle aside bar 라우팅 및 패키지 구조 이동, 리네이밍
